feat(cuda-ssh): Include `locales`
SSH passes the client's `LANG` and `LC_*` environment variables
by default. `bash` will complain if the locales specified in those
variables are not installed, but the only pre-installed locales on
most container images are C, C.UTF-8, and POSIX.

This adds the en_US.UTF-8 locale as well, and leaves `locale-gen`
available to install others so that more people can use the SSH server
to develop with their regular locale in effect more easily.
The other option is to disable the default `AcceptEnv` setting in
`/etc/ssh/sshd_config` to always use the container's default locale.
Both options have pros and cons, but it is more complicated to disable
the latter and switch to the former, so the default is the former.