fix(torch): Install the correct version of `libnccl2`
`libnccl2`'s installable package is versioned in the format
`libnccl2=2.A.B-C+cudaX.Y`, unlike other CUDA packages that
are versioned in their name, like `libcublas-X-Y=...`.
The install step for `libnccl2` then wasn't being narrowed down
to the correct CUDA version, so it would just install the version
for whatever the latest available CUDA was, rather than the version
present.
This change filters down the list of `libnccl2` version candidates
by the glob pattern `2.*+cudaX.Y`, using the appropriate major and minor
version for the ongoing build in order to install an
appropriate version of `libnccl2`.